feat(console/branches): refactor delete dialogue into it's own module
